pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "codenavigator"
version = "0.1.0"
description = "Code intelligence for any Python project — AST indexing, call graphs, dead code, impact analysis"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Code Generators",
    "Topic :: Utilities",
]

[project.optional-dependencies]
yaml   = ["pyyaml>=6.0"]
analysis = ["radon>=6.0", "vulture>=2.7"]
all    = ["pyyaml>=6.0", "radon>=6.0", "vulture>=2.7"]

[project.scripts]
codegrapher-index    = "codegrapher.core.indexer:main"
codegrapher-query    = "codegrapher.core.query_engine:main"
codegrapher-validate = "codegrapher.core.skill_validator:main"

[project.urls]
Homepage      = "https://github.com/vchandaria/codegrapher"
Repository    = "https://github.com/vchandaria/codegrapher.git"
Issues        = "https://github.com/vchandaria/codegrapher/issues"

[tool.setuptools.packages.find]
where = [".."]
include = ["codegrapher*"]

[tool.setuptools.package-data]
"codegrapher" = [
    "agents/*.md",
    "commands/*.md",
    "skills/**/*",
    ".claude-plugin/plugin.json",
]
